refactor(manager): DatabaseManager.csをRepository パターンで分割
1,763行の巨大ファイルを責務ごとに分割:
- DatabaseConnection.cs: 接続管理・WALモード・リトライロジック
- SchemaManager.cs: テーブル作成・マイグレーション(v1→v6)
- Repositories/GameRepository.cs: ゲームCRUD
- Repositories/VersionRepository.cs: バージョンCRUD
- Repositories/DeveloperRepository.cs: 開発者CRUD

DatabaseManager.csは薄いファサード(61行)として残し、既存の呼び出し元を壊さない。
